AutoHireAI/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "nltk>=3.9.1",
    "openai>=1.72.0",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "sendgrid>=6.11.0",
    "twilio>=9.5.2",
]
